❌ Remove deprecated calendar components Reduce calendar month to mini version only
